refactor: 메소드명, 래퍼타입 기본형으로 변경

+++ src/main/java/com/gio/ordersystem/orders/dto/OrderCreateRequest.java
@@ -5,22 +5,20 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class OrderCreateRequest {
 
     @NotNull(message = "사용자 ID는 필수입니다.")
-    private Long userId;
+    private long userId;
 
     @NotNull(message = "상품 ID는 필수입니다.")
-    private Long productId;
+    private long productId;
 
     @Min(value = 1, message = "주문 수량은 1 이상이어야 합니다.")
-    private Integer quantity;
+    private int quantity;
 
     @NotNull(message = "총 주문 금액은 필수입니다.")
-    private BigDecimal totalPrice;
+    private long totalPrice;
 
     public Order toEntity() {
         return Order.builder()
